+++ cloudant-client/src/main/java/com/cloudant/client/api/model/ReplicatorDocument.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 IBM Corp. All rights reserved.
+ * Copyright (c) 2015, 2021 IBM Corp. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy of the License at
@@ -16,7 +16,9 @@ package com.cloudant.client.api.model;
 
 import com.cloudant.client.org.lightcouch.Attachment;
 import com.cloudant.client.org.lightcouch.Replicator;
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -228,11 +230,19 @@ public class ReplicatorDocument {
     }
 
     public Integer getSinceSeq() {
-        return replicatorDocument.getSinceSeq();
+        return replicatorDocument.getSinceSeq().getAsInt();
+    }
+
+    public String getSinceSeqString() {
+        return replicatorDocument.getSinceSeq().getAsString();
     }
 
     public void setSinceSeq(Integer sinceSeq) {
-        replicatorDocument.setSinceSeq(sinceSeq);
+        replicatorDocument.setSinceSeq(new JsonParser().parse(sinceSeq.toString()));
+    }
+
+    public void setSinceSeq(String sinceSeq) {
+        replicatorDocument.setSinceSeq(new JsonParser().parse(sinceSeq));
     }
 
     public void setSourceIamApiKey(String iamApiKey) {