+++ cloudant-client/src/main/java/com/cloudant/client/org/lightcouch/Replicator.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 2011 lightcouch.org
- * Copyright (c) 2015, 2019 IBM Corp. All rights reserved.
+ * Copyright (c) 2015, 2021 IBM Corp. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy of the License at
@@ -274,7 +274,7 @@ public class Replicator {
         return this;
     }
 
-    public Replicator sinceSeq(Integer sinceSeq) {
+    public Replicator sinceSeq(JsonElement sinceSeq) {
         replicatorDoc.setSinceSeq(sinceSeq);
         return this;
     }